+++ src/components/Header.tsx
@@ -1,12 +1,14 @@
 import Box from "@mui/material/Box";
 import Divider from "@mui/material/Divider";
 import Typography from "@mui/material/Typography";
+import type {Theme} from "@mui/material/styles";
 import React from "react";
 
+import {WalletIcon} from "@unstoppabledomains/ui-components";
 import IconPlate from "@unstoppabledomains/ui-kit/icons/IconPlate";
-import UnstoppableWalletIcon from "@unstoppabledomains/ui-kit/icons/UnstoppableWalletIcon";
+import {makeStyles} from "@unstoppabledomains/ui-kit/styles";
 
-const styles = {
+const useStyles = makeStyles()((theme: Theme) => ({
   main: {
     padding: 2,
     marginTop: "-16px",
@@ -40,9 +42,9 @@ const styles = {
   },
   title2: {
     fontSize: "13px",
-    color: "#939799",
+    color: theme.palette.getContrastText(theme.palette.background.paper),
   },
-};
+}));
 
 interface Props {
   title?: string;
@@ -54,36 +56,40 @@ const Header: React.FC<Props> = ({
   iconPath,
   title = "Your Portal to Decentralized Web",
   subTitle = "This extension opens websites on the blockchain",
-}) => (
-  <Box sx={styles.main}>
-    <Box sx={styles.contentContainer}>
-      <Box sx={styles.logoContainer}>
-        {iconPath ? (
-          <Box
-            component="img"
-            sx={styles.logo}
-            src={iconPath}
-            alt="Unstoppable domains logo"
-          />
-        ) : (
-          <IconPlate size={50} variant="info">
-            <UnstoppableWalletIcon />
-          </IconPlate>
-        )}
-      </Box>
-      <Box sx={styles.descriptionContainer}>
-        <Box fontWeight="fontWeightBold">
-          <Typography noWrap variant="subtitle1" sx={styles.title1}>
-            {title}
+}) => {
+  const {classes} = useStyles();
+
+  return (
+    <Box className={classes.main}>
+      <Box className={classes.contentContainer}>
+        <Box className={classes.logoContainer}>
+          {iconPath ? (
+            <Box
+              component="img"
+              className={classes.logo}
+              src={iconPath}
+              alt="Unstoppable domains logo"
+            />
+          ) : (
+            <IconPlate size={50} variant="info">
+              <WalletIcon />
+            </IconPlate>
+          )}
+        </Box>
+        <Box className={classes.descriptionContainer}>
+          <Box fontWeight="fontWeightBold">
+            <Typography noWrap variant="subtitle1" className={classes.title1}>
+              {title}
+            </Typography>
+          </Box>
+          <Typography noWrap variant="subtitle2" className={classes.title2}>
+            {subTitle}
           </Typography>
         </Box>
-        <Typography noWrap variant="subtitle2" sx={styles.title2}>
-          {subTitle}
-        </Typography>
       </Box>
+      <Divider />
     </Box>
-    <Divider />
-  </Box>
-);
+  );
+};
 
 export default Header;