+++ src/pages/OnUpdated/index.tsx
@@ -1,52 +1,71 @@
+import Alert from "@mui/material/Alert";
 import Box from "@mui/material/Box";
 import Button from "@mui/material/Button";
 import Chip from "@mui/material/Chip";
 import Paper from "@mui/material/Paper";
 import Typography from "@mui/material/Typography";
 import Markdown from "markdown-to-jsx";
-import React, {useEffect} from "react";
+import React, {useState} from "react";
 import {useNavigate} from "react-router-dom";
+import useAsyncEffect from "use-async-effect";
 
 import {AppEnv} from "@unstoppabledomains/config";
 import {
+  WalletIcon,
+  getBootstrapState,
   useDomainConfig,
+  useFireblocksState,
   useTranslationContext,
 } from "@unstoppabledomains/ui-components";
+import IconPlate from "@unstoppabledomains/ui-kit/icons/IconPlate";
 
 import config from "../../config";
 import usePreferences from "../../hooks/usePreferences";
-import {getManifestVersion} from "../../lib/runtime";
+import {getManifestVersion, signOut} from "../../lib/runtime";
 import {setWalletPreferences} from "../../lib/wallet/preferences";
 import {useExtensionStyles} from "../../styles/extension.styles";
-import {DefaultPageView} from "../../types/wallet/preferences";
 
 const OnUpdated: React.FC = () => {
   const {classes, cx} = useExtensionStyles();
   const {setShowSuccessAnimation} = useDomainConfig();
+  const [walletState, setWalletState] = useFireblocksState();
+  const [isSignedOut, setIsSignedOut] = useState(false);
   const {preferences} = usePreferences();
   const [t] = useTranslationContext();
   const navigate = useNavigate();
 
-  useEffect(() => {
+  useAsyncEffect(async () => {
+    // show the upgrade animation
     setShowSuccessAnimation(true);
+
+    // sign out the user if specified by the config
+    if (config.SIGN_OUT_ON_UPDATE) {
+      const signInState = getBootstrapState(walletState);
+      if (signInState?.refreshToken) {
+        await signOut();
+        await setWalletState({});
+        setIsSignedOut(true);
+      }
+    }
   }, []);
 
-  const handleOptOut = async () => {
-    await handlePreference("legacy");
-  };
-
-  const handleUseWallet = async () => {
-    await handlePreference("wallet");
-  };
-
-  const handlePreference = async (view: DefaultPageView) => {
+  const handleClick = async () => {
+    // ensure preferences are populated
     if (!preferences) {
       return;
     }
 
-    preferences.DefaultView = view;
-    preferences.WalletEnabled = view === "wallet";
+    // save updated default preferences
+    preferences.DefaultView = "wallet";
+    preferences.WalletEnabled = true;
     preferences.VersionInfo = config.VERSION_DESCRIPTION;
+
+    // set existing wallet flag if signed out
+    if (isSignedOut) {
+      preferences.HasExistingWallet = true;
+    }
+
+    // save preferences and show the wallet
     await setWalletPreferences(preferences);
     navigate("/");
   };
@@ -57,8 +76,10 @@ const OnUpdated: React.FC = () => {
         <Box
           className={cx(classes.contentContainer, classes.fullHeightCentered)}
         >
-          <img src={chrome.runtime.getURL("/icon/browser.svg")} />
-          <Typography variant="h4" mt={1}>
+          <IconPlate size={85} variant="info">
+            <WalletIcon />
+          </IconPlate>
+          <Typography variant="h4" mt={2}>
             {t("extension.welcomeToVersion", {
               version: getManifestVersion() || "dev",
             })}
@@ -80,22 +101,16 @@ const OnUpdated: React.FC = () => {
           )}
         </Box>
         <Box className={classes.contentContainer}>
-          {preferences?.WalletEnabled ? (
-            <Button variant="contained" onClick={handleUseWallet} fullWidth>
-              {t("common.continue")}
-            </Button>
-          ) : (
-            <>
-              <Button variant="contained" onClick={handleUseWallet} fullWidth>
-                {t("extension.enableWallet")}
-              </Button>
-              <Box className={classes.contentContainer} mt={1}>
-                <Button variant="text" onClick={handleOptOut} size="small">
-                  {t("extension.continueWithoutWallet")}
-                </Button>
-              </Box>
-            </>
+          {isSignedOut && (
+            <Box className={classes.contentContainer} mb={1}>
+              <Alert variant="standard" severity="info" sx={{width: "100%"}}>
+                You were signed during this upgrade.
+              </Alert>
+            </Box>
           )}
+          <Button variant="contained" onClick={handleClick} fullWidth>
+            {t("wallet.letsGo")}
+          </Button>
         </Box>
       </Box>
     </Paper>