+++ src/pages/Wallet/SignInCta.tsx
@@ -6,47 +6,28 @@ import Typography from "@mui/material/Typography";
 import React from "react";
 
 import {AppEnv} from "@unstoppabledomains/config";
-import {useTranslationContext} from "@unstoppabledomains/ui-components";
+import {
+  WalletIcon,
+  useCustomTheme,
+  useTranslationContext,
+} from "@unstoppabledomains/ui-components";
 import IconPlate from "@unstoppabledomains/ui-kit/icons/IconPlate";
-import UnstoppableWalletIcon from "@unstoppabledomains/ui-kit/icons/UnstoppableWalletIcon";
 
 import config from "../../config";
-import {setBadgeCount} from "../../lib/runtime";
-import {openPopupWindow} from "../../scripts/liteWalletProvider/background";
 import {useExtensionStyles} from "../../styles/extension.styles";
 
 interface SignInCtaProps {
-  onSignInClick: () => Promise<void>;
+  onCreateWalletClicked: () => void;
+  onSignInClicked: () => void;
 }
 
-export const SignInCta: React.FC<SignInCtaProps> = ({onSignInClick}) => {
+export const SignInCta: React.FC<SignInCtaProps> = ({
+  onCreateWalletClicked,
+  onSignInClicked,
+}) => {
   const {classes, cx} = useExtensionStyles();
   const [t] = useTranslationContext();
-
-  const handleSignIn = async (newUser: boolean) => {
-    // handle the callback before opening the new window
-    await onSignInClick();
-
-    // find current window
-    const parentWindow = await chrome.windows.getLastFocused();
-    if (!parentWindow?.id) {
-      return;
-    }
-
-    // determine the popup URL
-    const popupUrl = chrome.runtime.getURL(
-      `index.html?request=${encodeURIComponent(JSON.stringify({type: "signInRequest", params: [newUser]}))}&parentWindowId=${parentWindow.id}#connect`,
-    );
-
-    // open the popup
-    await openPopupWindow(popupUrl, parentWindow.id);
-
-    // set a badge
-    await setBadgeCount(1);
-
-    // close the existing extension popup
-    chrome.extension.getViews({type: "popup"}).map(w => w.close());
-  };
+  const theme = useCustomTheme();
 
   return (
     <Paper className={classes.container}>
@@ -55,13 +36,13 @@ export const SignInCta: React.FC<SignInCtaProps> = ({onSignInClick}) => {
           className={cx(classes.contentContainer, classes.fullHeightCentered)}
         >
           <IconPlate size={100} variant="info">
-            <UnstoppableWalletIcon />
+            <WalletIcon />
           </IconPlate>
           <Typography variant="h4" mt={3}>
-            {t("wallet.title")}
+            {theme.wallet.title}
           </Typography>
           <Typography variant="body1" mt={1} mb={2}>
-            {t("manage.cryptoWalletDescriptionShort")}
+            {theme.wallet.subTitle}
           </Typography>
           {(config.NODE_ENV as AppEnv) !== "production" && (
             <Chip
@@ -77,17 +58,18 @@ export const SignInCta: React.FC<SignInCtaProps> = ({onSignInClick}) => {
             variant="contained"
             fullWidth
             className={classes.button}
-            onClick={() => handleSignIn(false)}
-          >
-            {t("wallet.beginSetup")}
-          </Button>
-          <Button
-            variant="outlined"
-            fullWidth
-            className={classes.button}
-            onClick={() => handleSignIn(true)}
+            onClick={onCreateWalletClicked}
           >
             {t("wallet.createWallet")}
+          </Button>
+          <Button
+            variant="text"
+            fullWidth
+            size="small"
+            className={classes.button}
+            onClick={onSignInClicked}
+          >
+            {t("wallet.alreadyHaveWallet")}
           </Button>
         </Box>
       </Box>